Improved intro to avoid white areas
Previously had problem where was using a mask on an area that
was only being cleared, and using "darken" mode. Turns out
pixels that are "cleared" are not set to level 0, but instead
to just a unset value that happens to appear as level 0. But
then white parts of the mask show up because the darkest pixel
was a white pixel on the mask.